refactor: rename serverTiming folder to server-timing
